Refactor MilkDrop compiler modules
