+++ assets/js/milkdrop/compiler/default-state.ts
@@ -1,0 +1,256 @@
+const MAX_CUSTOM_WAVES = 32;
+const MAX_CUSTOM_SHAPES = 32;
+
+function createDefaultShapeSlot(index: number): Record<string, number> {
+  if (index === 1) {
+    return {
+      shape_1_enabled: 1,
+      shape_1_sides: 6,
+      shape_1_x: 0.5,
+      shape_1_y: 0.5,
+      shape_1_rad: 0.17,
+      shape_1_ang: 0,
+      shape_1_a: 0.24,
+      shape_1_r: 1,
+      shape_1_g: 0.48,
+      shape_1_b: 0.84,
+      shape_1_a2: 0,
+      shape_1_r2: 0,
+      shape_1_g2: 0,
+      shape_1_b2: 0,
+      shape_1_border_a: 0.86,
+      shape_1_border_r: 1,
+      shape_1_border_g: 0.8,
+      shape_1_border_b: 1,
+      shape_1_additive: 1,
+      shape_1_thickoutline: 1,
+    };
+  }
+
+  const fallbackByIndex: Record<number, Record<string, number>> = {
+    2: {
+      sides: 5,
+      rad: 0.12,
+      a: 0.18,
+      r: 0.8,
+      g: 0.5,
+      b: 1,
+      border_a: 0.78,
+      border_r: 0.9,
+      border_g: 0.9,
+      border_b: 1,
+    },
+    3: {
+      sides: 4,
+      rad: 0.1,
+      a: 0.16,
+      r: 1,
+      g: 0.7,
+      b: 0.4,
+      border_a: 0.7,
+      border_r: 1,
+      border_g: 0.9,
+      border_b: 0.5,
+    },
+    4: {
+      sides: 8,
+      rad: 0.09,
+      a: 0.14,
+      r: 0.6,
+      g: 0.85,
+      b: 1,
+      border_a: 0.7,
+      border_r: 0.75,
+      border_g: 0.95,
+      border_b: 1,
+    },
+    5: {
+      sides: 7,
+      rad: 0.08,
+      a: 0.14,
+      r: 1,
+      g: 0.8,
+      b: 0.45,
+      border_a: 0.72,
+      border_r: 1,
+      border_g: 0.9,
+      border_b: 0.6,
+    },
+    6: {
+      sides: 3,
+      rad: 0.07,
+      a: 0.13,
+      r: 0.8,
+      g: 1,
+      b: 0.55,
+      border_a: 0.7,
+      border_r: 0.9,
+      border_g: 1,
+      border_b: 0.7,
+    },
+    7: {
+      sides: 9,
+      rad: 0.06,
+      a: 0.12,
+      r: 0.7,
+      g: 0.9,
+      b: 1,
+      border_a: 0.68,
+      border_r: 0.8,
+      border_g: 0.95,
+      border_b: 1,
+    },
+    8: {
+      sides: 5,
+      rad: 0.05,
+      a: 0.1,
+      r: 1,
+      g: 0.65,
+      b: 0.9,
+      border_a: 0.66,
+      border_r: 1,
+      border_g: 0.75,
+      border_b: 0.95,
+    },
+  };
+
+  const fallback = fallbackByIndex[index] ?? fallbackByIndex[8];
+  return {
+    [`shape_${index}_enabled`]: 0,
+    [`shape_${index}_sides`]: fallback.sides,
+    [`shape_${index}_x`]: 0.5,
+    [`shape_${index}_y`]: 0.5,
+    [`shape_${index}_rad`]: fallback.rad,
+    [`shape_${index}_ang`]: 0,
+    [`shape_${index}_a`]: fallback.a,
+    [`shape_${index}_r`]: fallback.r,
+    [`shape_${index}_g`]: fallback.g,
+    [`shape_${index}_b`]: fallback.b,
+    [`shape_${index}_a2`]: 0,
+    [`shape_${index}_r2`]: 0,
+    [`shape_${index}_g2`]: 0,
+    [`shape_${index}_b2`]: 0,
+    [`shape_${index}_border_a`]: fallback.border_a,
+    [`shape_${index}_border_r`]: fallback.border_r,
+    [`shape_${index}_border_g`]: fallback.border_g,
+    [`shape_${index}_border_b`]: fallback.border_b,
+    [`shape_${index}_additive`]: 0,
+    [`shape_${index}_thickoutline`]: 0,
+  };
+}
+
+function createDefaultCustomWaveSlot(index: number): Record<string, number> {
+  return {
+    [`custom_wave_${index}_enabled`]: 0,
+    [`custom_wave_${index}_samples`]: 64,
+    [`custom_wave_${index}_spectrum`]: 0,
+    [`custom_wave_${index}_additive`]: 0,
+    [`custom_wave_${index}_usedots`]: 0,
+    [`custom_wave_${index}_scaling`]: 1,
+    [`custom_wave_${index}_smoothing`]: 0.5,
+    [`custom_wave_${index}_mystery`]: 0,
+    [`custom_wave_${index}_thick`]: 1,
+    [`custom_wave_${index}_x`]: 0.5,
+    [`custom_wave_${index}_y`]: 0.5,
+    [`custom_wave_${index}_r`]: 1,
+    [`custom_wave_${index}_g`]: 1,
+    [`custom_wave_${index}_b`]: 1,
+    [`custom_wave_${index}_a`]: 0.4,
+  };
+}
+
+const DEFAULT_MILKDROP_STATE: Record<string, number> = {
+  fRating: 3,
+  beat_sensitivity: 0.7,
+  blend_duration: 2.4,
+  decay: 0.93,
+  zoom: 1,
+  zoomexp: 1,
+  rot: 0,
+  warp: 0.08,
+  cx: 0.5,
+  cy: 0.5,
+  sx: 1,
+  sy: 1,
+  dx: 0,
+  dy: 0,
+  warpanimspeed: 1,
+  shader: 1,
+  modwavealphastart: 1,
+  modwavealphaend: 1,
+  bmodwavealphabyvolume: 0,
+  wave_mode: 0,
+  wave_scale: 1,
+  wave_smoothing: 0.72,
+  wave_a: 0.86,
+  wave_r: 0.35,
+  wave_g: 0.7,
+  wave_b: 1,
+  wave_x: 0.5,
+  wave_y: 0.5,
+  wave_mystery: 0.18,
+  wave_thick: 1.4,
+  wave_additive: 1,
+  wave_usedots: 0,
+  wave_brighten: 1,
+  mesh_density: 18,
+  mesh_alpha: 0.22,
+  mesh_r: 0.3,
+  mesh_g: 0.5,
+  mesh_b: 0.95,
+  bg_r: 0.02,
+  bg_g: 0.03,
+  bg_b: 0.06,
+  brighten: 0,
+  darken: 0,
+  darken_center: 0,
+  solarize: 0,
+  invert: 0,
+  gammaadj: 1,
+  video_echo_enabled: 0,
+  video_echo_alpha: 0.18,
+  video_echo_zoom: 1.02,
+  video_echo_orientation: 0,
+  ob_size: 0,
+  ob_r: 0.92,
+  ob_g: 0.96,
+  ob_b: 1,
+  ob_a: 0.8,
+  ib_size: 0,
+  ib_r: 0.92,
+  ib_g: 0.96,
+  ib_b: 1,
+  ib_a: 0.76,
+  texture_wrap: 0,
+  feedback_texture: 0,
+  ob_border: 0,
+  ib_border: 0,
+  motion_vectors: 0,
+  motion_vectors_x: 16,
+  motion_vectors_y: 12,
+  mv_dx: 0,
+  mv_dy: 0,
+  mv_l: 0,
+  mv_r: 1,
+  mv_g: 1,
+  mv_b: 1,
+  mv_a: 0.35,
+  ...Object.fromEntries(
+    Array.from({ length: MAX_CUSTOM_SHAPES }, (_, index) =>
+      Object.entries(createDefaultShapeSlot(index + 1)),
+    ).flat(),
+  ),
+  ...Object.fromEntries(
+    Array.from({ length: MAX_CUSTOM_WAVES }, (_, index) =>
+      Object.entries(createDefaultCustomWaveSlot(index + 1)),
+    ).flat(),
+  ),
+};
+
+export {
+  createDefaultCustomWaveSlot,
+  createDefaultShapeSlot,
+  DEFAULT_MILKDROP_STATE,
+  MAX_CUSTOM_SHAPES,
+  MAX_CUSTOM_WAVES,
+};
